Tweak setup script so it doesn't fail

+++ setup.sh
@@ -6,7 +6,13 @@ git submodule update --init --recursive
 for name in *; do
   target="$HOME/.$name"
   if [ "$name" != 'setup.sh' ] && [ "$name" != 'README.md' ] && [ "$name" != 'LICENSE' ]; then
-    echo "Creating $target"
-    ln -s "$PWD/$name" "$target"
+    if [ -e "$target" ]; then
+      if [ ! -L "$target" ]; then
+        echo "WARNING: $target exists but is not a symlink."
+      fi
+    else
+      echo "Creating $target"
+      ln -s "$PWD/$name" "$target"
+    fi
   fi
 done
